Fix `Choices` Deserialization
Some OpenAPI Compatible APIs like Ollama don't always return `lobprobs`
as part of the `choices` JSON. To handle this we need to add a default
argument to `logprobs` so upickle can handle this case.

This PR moves the argument to the end which is breaking.

This PR also adds an ollama deserialization test for completions.